move to Benoit's new analyzer format. NOTE: Need to be extra careful about experiment naming now.
